+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+- Clarified the use of SVGs with a new test and changes to documentation. 
+
+## UNRELEASED
+
 - Introduced the `allowedEmptyAttributes` option, enabling explicit specification of empty string values for select attributes, with the default attribute set to `alt`.
 
 ## 2.11.0 (2023-06-21)